requirements.txt

langchain
langchain-openai
django
python-decouple